Distinguish between protocol level and UI level in LIP 0018
LIP 0018 is updated such that there is a clear distinction between addresses
on the protocol level (20 bytes) and the encoding for the UI level (adding a
checksum and using a base32 encoding).
Before, it was specified that the checksum and Base32 encoding has to be
validated during transaction validation. This is however meaningless since
neither the checksum nor any other encoding details are signed or hashed.
Only the 20 bytes of the public key hash are signed and hashed and therefore
relevant for the protocol.